+++ src/main/java/uk/gov/hmcts/reform/roleassignmentbatch/batch/BatchConfig.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.roleassignmentbatch1.batch;
+package uk.gov.hmcts.reform.roleassignmentbatch.batch;
 
 import org.springframework.batch.core.Job;
 import org.springframework.batch.core.Step;
@@ -11,7 +11,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import uk.gov.hmcts.reform.roleassignmentbatch1.task.DeleteExpiredRecords;
+import uk.gov.hmcts.reform.roleassignmentbatch.task.DeleteExpiredRecords;
 
 @Configuration
 @EnableBatchProcessing
