+++ src/smokeTest/java/uk/gov/hmcts/reform/roleassignmentbatch/EmptySmokeTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.roleassignmentbatch1;
+package uk.gov.hmcts.reform.roleassignmentbatch;
 
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
