+++ src/test/java/uk/gov/hmcts/reform/roleassignmentbatch/RoleAssignmentBatchApplicationTests.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.roleassignmentbatch1;
+package uk.gov.hmcts.reform.roleassignmentbatch;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
